Reorganized + Minor fixes
