+++ .github/workflows/ci.yml
@@ -67,15 +67,21 @@ jobs:
 
       - name: Check out code into the Go module directory
         uses: actions/checkout@v3
+      
+      - name: Install Kind
+        # Before updating the kind version to use, verify that the current kind image
+        # is still supported by the version.
+        run: |
+          go install sigs.k8s.io/kind@v0.20.0
 
       - name: Run e2e tests
         run: |
-          OUTPUT_TYPE=type=docker make docker-build-member-agent docker-build-hub-agent docker-build-refresh-token e2e-tests
+          OUTPUT_TYPE=type=docker make docker-build-member-agent docker-build-hub-agent docker-build-refresh-token e2e-tests-v1alpha1
         env:
           KUBECONFIG: '/home/runner/.kube/config'
           HUB_SERVER_URL: 'https://172.19.0.2:6443'
   
-  e2e-tests-v1beta1:
+  e2e-tests:
     runs-on: ubuntu-latest
     needs: [
       detect-noop,
@@ -93,9 +99,15 @@ jobs:
       - name: Install Ginkgo CLI
         run: |
           go install github.com/onsi/ginkgo/v2/ginkgo@v2.9.5
+      
+      - name: Install Kind
+        # Before updating the kind version to use, verify that the current kind image
+        # is still supported by the version.
+        run: |
+          go install sigs.k8s.io/kind@v0.20.0
 
       - name: Run e2e tests
         run: |
-          make e2e-tests-v1beta1
+          make e2e-tests
         env:
           KUBECONFIG: '/home/runner/.kube/config'
