+++ test/e2e/utils_test.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package v1beta1
+package e2e
 
 import (
 	"fmt"
